requirements.txt
numpy==2.2.1
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
six==1.17.0
tzdata==2024.2
